fix: ec2 builder validation step was not working (#325)
*Issue #, if available:*

*Description of changes:*
The validation step was not working properly; it unexpectedly succeed
even when some of the commands are missing. Thi s PR adds `|| exit 1`
for every command to exit with error when a command failed.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.